Module 12 Small refactoring
